% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treemap.R
\name{renderTreemap}
\alias{renderTreemap}
\title{Widget render function for use in Shiny}
\usage{
renderTreemap(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

