% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treemap.R
\name{treemapOutput}
\alias{treemapOutput}
\title{Widget output function for use in Shiny}
\usage{
treemapOutput(outputId, width = "600px", height = "600px")
}
\description{
Widget output function for use in Shiny
}

